Fix warnings on 32 bit sim.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24347 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/aiff.c
@@ -218,7 +218,8 @@ next_track:
     codec = get_codec(format.formattag);
     if (codec == 0)
     {
-        DEBUGF("CODEC_ERROR: AIFC does not support compressionType: 0x%x\n", format.formattag);
+        DEBUGF("CODEC_ERROR: AIFC does not support compressionType: 0x%x\n", 
+            (unsigned int)format.formattag);
         i = CODEC_ERROR;
         goto done;
     }
